+++ src/call_client/event_handler.rs
@@ -57,6 +57,39 @@ impl PyEventHandler {
         Ok(())
     }
 
+    /// Event emitted when the session with the dial-out remote end is
+    /// established.
+    ///
+    /// :param Mapping[str, Any] data: See :ref:`DailoutEvent`
+    fn on_dialout_connected(&self, data: PyObject) -> PyResult<()> {
+        Ok(())
+    }
+
+    /// Event emitted in the case of dial-out errors which are fatal and the
+    /// service cannot proceed. For example, an error in SDP negotiation is
+    /// fatal to the media/SIP pipeline and will result in dialout-error being
+    /// triggered.
+    ///
+    /// :param Mapping[str, Any] data: See :ref:`DailoutEvent`
+    fn on_dialout_error(&self, data: PyObject) -> PyResult<()> {
+        Ok(())
+    }
+
+    /// Event emitted when the dial-out remote end disconnects the call or the
+    /// call is stopped by calling :ref:`daily.CallClient.stop_dialout`.
+    ///
+    /// :param Mapping[str, Any] data: See :ref:`DailoutEvent`
+    fn on_dialout_stopped(&self, data: PyObject) -> PyResult<()> {
+        Ok(())
+    }
+
+    /// Event emitted when a dial-out warning occurs.
+    ///
+    /// :param Mapping[str, Any] data: See :ref:`DailoutEvent`
+    fn on_dialout_warning(&self, data: PyObject) -> PyResult<()> {
+        Ok(())
+    }
+
     /// Event emitted when an error occurs.
     ///
     /// :param string message: The error message