+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to the **daily-python** SDK will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added Dial-Out events `dialout-connected`, `dialout-stopped`, `dialout-error`
+  and `dialout-warning`.
+
 ## [0.7.4] - 2024-04-16
 
 ### Fixed